Warn when the used toolchain looks too old for rust-analyzer
