+++ crates/rust-analyzer/src/reload.rs
@@ -182,6 +182,22 @@ impl GlobalState {
                 self.proc_macro_clients.iter().map(Some).chain(iter::repeat_with(|| None));
 
             for (ws, proc_macro_client) in self.workspaces.iter().zip(proc_macro_clients) {
+                if let Some(toolchain) = &ws.toolchain {
+                    if *toolchain < crate::MINIMUM_SUPPORTED_TOOLCHAIN_VERSION {
+                        status.health |= lsp_ext::Health::Warning;
+                        format_to!(
+                            message,
+                            "Workspace `{}` is using an outdated toolchain version `{}` but \
+                            rust-analyzer only supports `{}` and higher.\n\
+                            Consider using the rust-analyzer rustup component for your toolchain or
+                            upgrade your toolchain to a supported version.\n\n",
+                            ws.manifest_or_root(),
+                            toolchain,
+                            crate::MINIMUM_SUPPORTED_TOOLCHAIN_VERSION,
+                        );
+                    }
+                }
+
                 if let ProjectWorkspaceKind::Cargo { error: Some(error), .. }
                 | ProjectWorkspaceKind::DetachedFile {
                     cargo: Some((_, _, Some(error))), ..
